io_uring_prep_cmd_sock: cast pointer to untptr_t first
For 32-bit archs, cast it to an uintptr_t first to avoid warnings.

Fixes: 3797db16928a ("liburing.h: Populate SQE for {s,g} etsockopt")

+++ src/include/liburing.h
@@ -1151,7 +1151,7 @@ IOURINGINLINE void io_uring_prep_cmd_sock(struct io_uring_sqe *sqe,
 					  int optlen)
 {
 	io_uring_prep_rw(IORING_OP_URING_CMD, sqe, fd, NULL, 0, 0);
-	sqe->optval = (long long unsigned int)optval;
+	sqe->optval = (unsigned long) (uintptr_t) optval;
 	sqe->optname = optname;
 	sqe->optlen = optlen;
 	sqe->cmd_op = cmd_op;
